+++ packages/npm/less-plugins/packages/shared/src/constants.ts
@@ -1,0 +1,1 @@
+export const FILE_NAME_PACKAGE_JSON = 'package.json';