+++ packages/npm/less-plugins/packages/shared/src/parse-options.ts
@@ -1,0 +1,5 @@
+import yargs from 'yargs';
+
+export function parseOptions(options: string) {
+    return yargs(options).parse();
+}