+++ packages/npm/less-plugins/packages/playground/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.9
+
+- Update dependencies @less-plugins/get-hashed-version@0.0.7
+
+- Update dependencies @less-plugins/shared@0.0.6
+
 ## 0.0.8
 
 - Update dependencies @less-plugins/rtl@0.0.6