+++ packages/npm/less-plugins/packages/replace-properties/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.5
+
+- Update dependencies @less-plugins/shared@0.0.6
+
 ## 0.0.4
 
 - update replace reg