+++ packages/npm/less-plugins/packages/rtl/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.7
+
+- Update dependencies @less-plugins/shared@0.0.6
+
+- Update dependencies @less-plugins/css-logical@0.0.8
+
+- Update dependencies @less-plugins/replace-properties@0.0.5
+
 ## 0.0.6
 
 - update rtl mode